+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Features
 
 - New Sqlite instrumentation ([#502](https://github.com/getsentry/sentry-android-gradle-plugin/pull/502))
+    - This integration replaces the old database instrumentation with the `sentry-android-sqlite` integration
+    - Any implementation of SupportSQLiteOpenHelper.Factory is now supported
 
 ### Fixes